+++ crates/bevy_pbr/src/material_bind_groups.rs
@@ -11,7 +11,6 @@ use bevy_ecs::{
     system::{Commands, Res},
 };
 use bevy_platform::collections::{HashMap, HashSet};
-use bevy_reflect::{prelude::ReflectDefault, Reflect};
 use bevy_render::render_resource::{BindlessSlabResourceLimit, PipelineCache};
 use bevy_render::{
     render_resource::{
@@ -31,6 +30,7 @@ use bytemuck::Pod;
 use core::hash::Hash;
 use core::{cmp::Ordering, iter, mem, ops::Range};
 use tracing::{error, trace};
+pub use bevy_render::mesh::material_bind_group::*;
 
 #[derive(Resource, Deref, DerefMut, Default)]
 pub struct MaterialBindGroupAllocators(TypeIdMap<MaterialBindGroupAllocator>);
@@ -253,43 +253,6 @@ enum BindingResourceArray<'a> {
     Samplers(Vec<&'a WgpuSampler>),
 }
 
-/// The location of a material (either bindless or non-bindless) within the
-/// slabs.
-#[derive(Clone, Copy, Debug, Default, Reflect)]
-#[reflect(Clone, Default)]
-pub struct MaterialBindingId {
-    /// The index of the bind group (slab) where the GPU data is located.
-    pub group: MaterialBindGroupIndex,
-    /// The slot within that bind group.
-    ///
-    /// Non-bindless materials will always have a slot of 0.
-    pub slot: MaterialBindGroupSlot,
-}
-
-/// The index of each material bind group.
-///
-/// In bindless mode, each bind group contains multiple materials. In
-/// non-bindless mode, each bind group contains only one material.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, Reflect, Deref, DerefMut)]
-#[reflect(Default, Clone, PartialEq, Hash)]
-pub struct MaterialBindGroupIndex(pub u32);
-
-impl From<u32> for MaterialBindGroupIndex {
-    fn from(value: u32) -> Self {
-        MaterialBindGroupIndex(value)
-    }
-}
-
-/// The index of the slot containing material data within each material bind
-/// group.
-///
-/// In bindless mode, this slot is needed to locate the material data in each
-/// bind group, since multiple materials are packed into a single slab. In
-/// non-bindless mode, this slot is always 0.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Reflect, Deref, DerefMut)]
-#[reflect(Default, Clone, PartialEq)]
-pub struct MaterialBindGroupSlot(pub u32);
-
 /// The CPU/GPU synchronization state of a buffer that we maintain.
 ///
 /// Currently, the only buffer that we maintain is the
@@ -397,18 +360,6 @@ where
 /// This is essentially arbitrary, as it doesn't seem to matter to `wgpu` what
 /// the size is.
 const DEFAULT_BINDLESS_FALLBACK_BUFFER_SIZE: u64 = 16;
-
-impl From<u32> for MaterialBindGroupSlot {
-    fn from(value: u32) -> Self {
-        MaterialBindGroupSlot(value)
-    }
-}
-
-impl From<MaterialBindGroupSlot> for u32 {
-    fn from(value: MaterialBindGroupSlot) -> Self {
-        value.0
-    }
-}
 
 impl<'a> From<&'a OwnedBindingResource> for BindingResourceId {
     fn from(value: &'a OwnedBindingResource) -> Self {
